+++ src/args.rs
@@ -1,127 +1,55 @@
-use std::{path::PathBuf, process::Command};
+use std::path::PathBuf;
 
 use clap::Parser;
-use comfy_table::Table;
-use log::debug;
 
 use crate::{
-    data::{Project, Size},
-    total_size_of_projects,
+    data::Size,
+    utils::{self, total_size_of_projects},
 };
 
-/// Represents the command line arguments
+/// Represents the command line options
 #[derive(Parser)]
-#[command(author, version, about)]
-pub struct Args {
+pub struct Opts {
     /// Path to a directory
     #[clap(long, short, default_value = ".")]
     pub path: PathBuf,
     /// Run the program without making any changes
     #[clap(short, long)]
     pub dry_run: bool,
-    /// Command to execute
-    #[clap(subcommand)]
-    pub command: Commander,
 }
 
 /// Represents the available commands
 #[derive(Parser)]
+#[command(author, version, about)]
 pub enum Commander {
     /// Print statistics about all projects
-    Stats,
+    Stats(Opts),
     /// Calculate the total size of all target folders
-    Size,
+    Size(Opts),
     /// Execute `cargo clean` on all projects
-    Clean,
+    Clean(Opts),
 }
 
 impl Commander {
-    pub fn show(&self, projects: &[Project], dry_run: bool) -> anyhow::Result<()> {
+    pub fn show(&self) -> anyhow::Result<()> {
         match self {
-            Self::Stats => {
-                Self::show_stats(projects);
+            Self::Stats(opts) => {
+                let (projects, _) = utils::check_args(opts)?;
+                utils::show_stats(&projects);
             }
-            Self::Size => {
+            Self::Size(opts) => {
+                let (projects, _) = utils::check_args(opts)?;
                 println!(
                     "Total size: {} ({} Projects)",
-                    Size::to_size(total_size_of_projects(projects)),
+                    Size::to_size(total_size_of_projects(&projects)),
                     projects.len()
                 );
             }
-            Self::Clean => {
-                Self::run_clean(projects, dry_run)?;
+            Self::Clean(opts) => {
+                let (projects, dry_run) = utils::check_args(opts)?;
+                utils::run_clean(&projects, dry_run)?;
             }
         }
-
-        Ok(())
-    }
-
-    fn show_stats(projects: &[Project]) {
-        let mut table = Table::new();
-        table.set_header(vec!["Project", "Size", "Path"]);
-        for project in projects {
-            table.add_row(vec![
-                &project.name,
-                &project.size.to_string(),
-                &project.path.to_string_lossy().to_string(),
-            ]);
-        }
-
-        table.add_row(vec![
-            "Total",
-            &Size::to_size(total_size_of_projects(projects)).to_string(),
-        ]);
-        print!("{table}");
-    }
-
-    fn print_status(projects: &[Project], cleaned: &[Project]) {
-        println!(
-            "Cleaned up: {} ({} Projects)\n {}",
-            Size::to_size(total_size_of_projects(cleaned)),
-            projects.len(),
-            cleaned
-                .iter()
-                .map(|p| p.name.clone())
-                .collect::<Vec<String>>()
-                .join(", ")
-        );
-    }
-
-    fn run_clean(projects: &[Project], dry_run: bool) -> anyhow::Result<()> {
-        let mut cleaned_projects = vec![];
-
-        for project in projects {
-            let target_path = project.path.join("target");
-
-            if dry_run {
-                debug!("Would remove: {:?}", target_path);
-            } else {
-                debug!("Running `cargo clean` for project: {:?}", project.name);
-
-                let result = Command::new("cargo")
-                    .arg("clean")
-                    .current_dir(&project.path)
-                    .output();
-
-                match result {
-                    Ok(output) => {
-                        if output.status.success() {
-                            cleaned_projects.push(project.clone());
-                        } else {
-                            let error_message = String::from_utf8_lossy(&output.stderr).to_string();
-                            Self::print_status(projects, &cleaned_projects);
-                            anyhow::bail!("Failed to clean {}: {}", project.name, error_message);
-                        }
-                    }
-                    Err(e) => {
-                        Self::print_status(projects, &cleaned_projects);
-                        anyhow::bail!("Failed to clean {}: {}", project.name, e);
-                    }
-                }
-            }
-        }
-
-        Self::print_status(projects, &cleaned_projects);
 
         Ok(())
     }
